// client/src/shared/middleware/platform/configuration/models/Workflow.ts
export interface Workflow {
    /**
     * The created by.
     * @type {string}
     * @memberof Workflow
     */
    readonly createdBy?: string;
    /**
     * The created date.
     * @type {Date}
     * @memberof Workflow
     */
    readonly createdDate?: Date;
    /**
     * The number of workflow connections
     * @type {number}
     * @memberof Workflow
     */
    readonly connectionsCount?: number;
    /**
     * The definition of a workflow.
     * @type {string}
     * @memberof Workflow
     */
    definition?: string;
    /**
     * The description of a workflow.
     * @type {string}
     * @memberof Workflow
     */
    description?: string;
    /**
     * 
     * @type {WorkflowFormat}
     * @memberof Workflow
     */
    format?: WorkflowFormat;
    /**
     * The id of a workflow.
     * @type {string}
     * @memberof Workflow
     */
    readonly id?: string;
    /**
     * The workflow's expected list of inputs.
     * @type {Array<WorkflowInput>}
     * @memberof Workflow
     */
    readonly inputs?: Array<WorkflowInput>;
    /**
     * The number of workflow inputs
     * @type {number}
     * @memberof Workflow
     */
    readonly inputsCount?: number;
    /**
     * The descriptive name for the workflow
     * @type {string}
     * @memberof Workflow
     */
    readonly label?: string;
    /**
     * The last modified by.
     * @type {string}
     * @memberof Workflow
     */
    readonly lastModifiedBy?: string;
    /**
     * The last modified date.
     * @type {Date}
     * @memberof Workflow
     */
    readonly lastModifiedDate?: Date;
    /**
     * The workflow's list of expected outputs.
     * @type {Array<WorkflowOutput>}
     * @memberof Workflow
     */
    readonly outputs?: Array<WorkflowOutput>;
    /**
     * The type of the source which stores the workflow definition.
     * @type {WorkflowSourceTypeEnum}
     * @memberof Workflow
     */
    sourceType?: WorkflowSourceTypeEnum;
    /**
     * The maximum number of times a task may retry.
     * @type {number}
     * @memberof Workflow
     */
    readonly maxRetries?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof Workflow
     */
    readonly workflowTaskComponentNames?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof Workflow
     */
    readonly workflowTriggerComponentNames?: Array<string>;
    /**
     * The steps that make up the workflow.
     * @type {Array<WorkflowTask>}
     * @memberof Workflow
     */
    readonly tasks?: Array<WorkflowTask>;
    /**
     * The steps that make up the workflow.
     * @type {Array<WorkflowTrigger>}
     * @memberof Workflow
     */
    readonly triggers?: Array<WorkflowTrigger>;
    /**
     * 
     * @type {number}
     * @memberof Workflow
     */
    version?: number;
}


/**
 * @export
 */
export const WorkflowSourceTypeEnum = {
    Classpath: 'CLASSPATH',
    Filesystem: 'FILESYSTEM',
    Git: 'GIT',
    Jdbc: 'JDBC'
} as const;
export type WorkflowSourceTypeEnum = typeof WorkflowSourceTypeEnum[keyof typeof WorkflowSourceTypeEnum];